move util to proper file
