+++ common/utils.go
@@ -4,6 +4,10 @@ package common
 
 import (
 	"regexp"
+	"unicode"
+	"unicode/utf8"
+
+	"github.com/rivo/uniseg"
 )
 
 var usernameRegex *regexp.Regexp = regexp.MustCompile(`^[0-9a-zA-Z_-]*[a-zA-Z0-9]+[0-9a-zA-Z_-]*$`)
@@ -15,4 +19,18 @@ const InvalidNameError string = `Invalid name.<br />Name must be between 3 and 3
 func IsValidName(name string) bool {
 	return 3 <= len(name) && len(name) <= 36 &&
 		usernameRegex.MatchString(name)
+}
+
+// TODO: There must be a library that can do this better
+func IsValidEmoji(s string) bool {
+	if uniseg.GraphemeClusterCount(s) != 1 {
+		return false
+	}
+
+	r, _ := utf8.DecodeRuneInString(s)
+	if r == utf8.RuneError {
+		return false
+	}
+
+	return !unicode.IsLetter(r) && !unicode.IsDigit(r)
 }
